% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generation.R
\name{FBA4Greatmod.generation}
\alias{FBA4Greatmod.generation}
\title{Flux Balance Analysis model generation for GreatMod}
\usage{
FBA4Greatmod.generation(
  fba_mat = NULL,
  fba_model = NULL,
  S = NULL,
  ub = NULL,
  lb = NULL,
  obj_fun = NULL,
  react_name = NULL
)
}
\arguments{
\item{fba_mat}{Matlab file .....}

\item{fba_model}{List representing the FBA model generated from \emph{FBAfile.generation}.}

\item{S}{Stoic matrix}

\item{ub}{Vector of the upper constraint of the fluxes}

\item{lb}{Vector of the lower constraint of the fluxes}

\item{obj_fun}{Objective function to maximize. Vector of zeros with length equal to the number of reactions,
with one 1 in the position of the flux to maximize.}

\item{react_name}{The reactions name}
}
\description{
....
}
\details{
Problem: S\emph{x = b -. size x = n and size of S = m X n
With this function the file to pass to the GLPKsolver is generated:
1st row) n_row n_col GLP_MAX (or GLP_MIN if the obj has to max or min)
2nd row) the coeff for defining the obj function (the number of coeff has to be == length of x)
m rows)   definition of the S row bounds (i.e. b)
n rows)   definition of the x bounds
m}n rows) definition of the S coeffs: row_index col_index value
}
\author{
Aucello Riccardo, Beccuti Marco, Cordero Francesca, Pernice Simone
}
