% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBAgreatmodeClass.R
\name{FBA_greatmod}
\alias{FBA_greatmod}
\title{Create an FBA_greatmod Object}
\usage{
FBA_greatmod(
  S,
  ub,
  lb,
  obj_fun,
  react_name = NULL,
  met_name = NULL,
  bioMax = -1,
  bioMean = -1,
  bioMin = -1,
  gene_assoc = NULL,
  pFBAFlag = -1
)
}
\value{
An object of class \code{FBA_greatmod}
}
\description{
Constructs a new object of the \code{FBA_greatmod} S4 class using model inputs.
}
