% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMat.R
\name{FBAmat.read}
\alias{FBAmat.read}
\title{Read matlab model}
\usage{
FBAmat.read(fba_mat)
}
\arguments{
\item{fba_mat}{Full path to matlab model file}
}
\description{
The generic function \code{readMATmod} imports matlab cobra (ONLY) models into ....
}
\details{
Returns ...
}
\author{
Aucello Riccardo, Beccuti Marco, Cordero Francesca, Pernice Simone
}
