% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMatGUI.R
\name{FBAmat.readGUI}
\alias{FBAmat.readGUI}
\title{Read matlab model}
\usage{
FBAmat.readGUI(fba_mat, input_dir)
}
\arguments{
\item{fba_mat}{Full path to matlab model file}
}
\description{
The generic function \code{readMATmod} imports matlab cobra (ONLY) models into ....
}
\details{
Returns ...
}
