% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pnpro.R
\name{generate_pnpro}
\alias{generate_pnpro}
\title{Generate a PNPRO XML from an arc table – DEBUG build}
\usage{
generate_pnpro(arc_df, pnpro_out)
}
\arguments{
\item{arc_df}{Data-frame containing columns
\verb{place, transition, direction, multiplicity, command}}

\item{pnpro_out}{Path to write the \emph{.PNPRO} file.}
}
\description{
Generate a PNPRO XML from an arc table – DEBUG build
}
