% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_R_from_pnpro.R
\name{generate_R_from_pnpro}
\alias{generate_R_from_pnpro}
\title{Generate a hypernode‐specific R script from:
• a .PNPRO file (to get place ordering)
• the functions_hypernode_template.R (with two stubs)
• your biounit_models list (to get cell & biomass initials)}
\usage{
generate_R_from_pnpro(
  pnpro_file,
  r_template,
  biounit_models,
  output_r = r_template
)
}
\arguments{
\item{pnpro_file}{Path to the .PNPRO XML file}

\item{r_template}{Path to functions_hypernode_template.R}

\item{biounit_models}{List of model objects (each with $initial_count and $biomass$mean)}

\item{output_r}{Path for the generated R script (defaults to overwrite r_template)}
}
\value{
Invisibly, the character vector of lines written
}
\description{
Generate a hypernode‐specific R script from:
• a .PNPRO file (to get place ordering)
• the functions_hypernode_template.R (with two stubs)
• your biounit_models list (to get cell & biomass initials)
}
