% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis_GUI.R
\name{model_analysis_GUI}
\alias{model_analysis_GUI}
\title{General wrapper for epimod::model.analysis (GUI) with console debug}
\usage{
model_analysis_GUI(
  paths,
  hypernode_name,
  debug_solver = FALSE,
  i_time = 0,
  f_time = 10,
  s_time = 1,
  atol = 1e-06,
  rtol = 1e-06,
  fba_fname,
  user_files = character(),
  volume = getwd()
)
}
\description{
General wrapper for epimod::model.analysis (GUI) with console debug
}
