% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_bounds_module.R
\name{model_dir}
\alias{model_dir}
\title{Strip the directory and “.mat” extension from a path.}
\usage{
model_dir(path)
}
\description{
Strip the directory and “.mat” extension from a path.
}
\keyword{internal}
