% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_GUI.R
\name{plot_metabolite_dynamics}
\alias{plot_metabolite_dynamics}
\title{Plot Extracellular Metabolite Concentrations Over Time}
\usage{
plot_metabolite_dynamics(hypernode_dir, output_dir = NULL)
}
\arguments{
\item{hypernode_dir}{Character. Path to the hypernode directory.}

\item{output_dir}{Character. Where to save plots (if NULL, does not save).}
}
\value{
A ggplot object of metabolite concentration dynamics.
}
\description{
Plot Extracellular Metabolite Concentrations Over Time
}
