% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_GUI.R
\name{plot_species_dynamics}
\alias{plot_species_dynamics}
\title{Plot Species Abundance and Biomass Dynamics for a Hypernode}
\usage{
plot_species_dynamics(hypernode_dir, final_time = Inf, output_dir = NULL)
}
\arguments{
\item{hypernode_dir}{Character. Path to the top-level hypernode directory.}

\item{final_time}{Numeric. Maximum time to plot (default: Inf).}

\item{output_dir}{Character. Where to save plots (if NULL, does not save).}
}
\value{
A ggplot object of species abundance over time.
}
\description{
Plot Species Abundance and Biomass Dynamics for a Hypernode
}
