% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{process_model}
\alias{process_model}
\title{Process a single biounit model}
\usage{
process_model(m, hypernode_name, base_dir = getwd())
}
\arguments{
\item{m}{list returned by make_biounit_models()}

\item{hypernode_name}{character, parent hyper-node label}

\item{base_dir}{where hypernodes/, metabolic_networks_library/ …
live (default = getwd()).}
}
\value{
list with status + paths (invisibly)
}
\description{
Process a single biounit model
}
