% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cobra_safe.R
\name{read_cobra_safe}
\alias{read_cobra_safe}
\title{Lettura sicura di file MAT (C++ nativo se presente, altrimenti fallback JSON)}
\usage{
read_cobra_safe(path)
}
\arguments{
\item{path}{Percorso al file \code{.mat}}
}
\value{
Lista con campi \code{S}, \code{lb}, \code{ub}, \code{c}, \code{rxns}, \code{mets}
}
\description{
Lettura sicura di file MAT (C++ nativo se presente, altrimenti fallback JSON)
}
