% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mode_json.R
\name{read_model_json}
\alias{read_model_json}
\title{Read a COBRA model JSON into the same list structure as read_cobra_mat()}
\usage{
read_model_json(json_path)
}
\arguments{
\item{json_path}{Path to the JSON file produced by mat2json.py}
}
\value{
A named list with the \emph{old} MAT field names:
S, lb, ub, c, rxns, rxnNames, mets, metNames, subSystems, rev,
rxnGeneMat, genes, grRules, rxnKEGGID, metKEGGID
}
\description{
Read a COBRA model JSON into the same list structure as read_cobra_mat()
}
