% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_FBA_R.R
\name{SA_FBA}
\alias{SA_FBA}
\title{Perform Sensitivity Analysis Using Flux Balance Analysis (FBA)}
\usage{
SA_FBA(
  result_dir,
  model_file,
  all_react_file,
  param_values_file,
  bounds_set_file,
  fbasol_t_file,
  fbasol_file,
  Y_file,
  Y2_file,
  indices_file,
  Y2_reaction,
  N,
  cores,
  b1,
  b2
)
}
\arguments{
\item{result_dir}{Directory to save results and intermediate files.}

\item{model_file}{Path to the metabolic model file.}

\item{all_react_file}{RDS file containing reaction annotations.}

\item{param_values_file}{File path to save the sampled parameter matrix.}

\item{bounds_set_file}{RDS file containing bounds for parameters.}

\item{fbasol_t_file}{Path to temporary FBA solution file.}

\item{fbasol_file}{Path to final FBA solution file.}

\item{Y_file}{File to save outputs from first objective.}

\item{Y2_file}{File to save outputs from second objective.}

\item{indices_file}{File to save computed sensitivity indices.}

\item{Y2_reaction}{ID of the reaction associated with second objective.}

\item{N}{Number of samples (e.g., 2^13).}

\item{cores}{Number of cores for parallel computing.}

\item{b1}{Lower bound for sampling parameters.}

\item{b2}{Upper bound for sampling parameters.}
}
\value{
Saves RDS files of sensitivity indices, parameter values, and FBA solutions.
}
\description{
This function performs a global sensitivity analysis using FBA simulations.
It evaluates the impact of model parameter variations on a predefined reaction objective.
The process involves Sobol' sampling, FBA runs in parallel, and computation of Sobol' indices
with bootstrapped confidence intervals.
}
\examples{
\dontrun{
SA_FBA(result_dir = "results/",
       model_file = "model.sbml",
       all_react_file = "reactions.rds",
       param_values_file = "params.rds",
       bounds_set_file = "bounds.rds",
       fbasol_t_file = "tmp_sol.rds",
       fbasol_file = "sol.rds",
       Y_file = "Y.rds",
       Y2_file = "Y2.rds",
       indices_file = "indices.rds",
       Y2_reaction = "biomass205",
       N = 8192,
       cores = 4,
       b1 = -10.0,
       b2 = 0.0)
}
}
