% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBAgreatmodeClass.R
\name{setObjFun}
\alias{setObjFun}
\alias{setObjFun,FBA_greatmod-method}
\title{Set Objective Function for FBA Model}
\usage{
setObjFun(theObject, reaction_name)
}
\arguments{
\item{theObject}{An object of class \code{FBA_greatmod}}

\item{reaction_name}{A character vector of reaction names to set as objective}
}
\value{
setObjFun returns the updated \code{FBA_greatmod} object with modified objective coefficients
}
\description{
This method assigns the objective coefficients based on the specified reaction names.
}
\examples{
\dontrun{
model <- setObjFun(model, reaction_name = "EX_glc__D_e")
}
}
