% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pnpro.R
\name{validate_pnpro}
\alias{validate_pnpro}
\title{Validate and repair a PNPRO file against a set of FBA models}
\usage{
validate_pnpro(
  pnpro2validate,
  hypernode_root,
  biounit_models,
  boundary_metabolites,
  out_dir,
  hypernode_name
)
}
\description{
Validate and repair a PNPRO file against a set of FBA models
}
