% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBAgreatmodeClass.R
\docType{methods}
\name{FBA_greatmod-methods}
\alias{FBA_greatmod-methods}
\alias{setDiet}
\alias{setDiet.name}
\alias{writeFBAfile}
\title{FBA_greatmod methods.}
\usage{
setDiet(theObject, dietf)

setDiet.name(theObject, diet_name)

writeFBAfile(theObject, fba_fname, ...)
}
\arguments{
\item{theObject}{A `FBA_greatmod` object}

\item{dietf}{The diet file or data.frame}

\item{diet_name}{Character for the diet name}

\item{fba_fname}{The file name in which the FBA model will be saved. See *epimod::model.generation*.}
}
\value{
The FBA_greatmod class with lower bounds updated with diet values.

The FBA_greatmod class with the diet name updated.
}
\description{
Write the fba model in file.
}
