% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBAgreatmodeClass.R
\docType{methods}
\name{getExchangesR}
\alias{getExchangesR}
\alias{FBA_greatmod-methods}
\alias{getConstraints}
\alias{setConstraints}
\alias{setDiet}
\alias{setDiet.name}
\alias{writeFBAfile}
\title{FBA_greatmod methods.}
\usage{
getExchangesR(theObject, reaction.name)

getConstraints(theObject, reaction.name)

setConstraints(theObject, reaction.name, newConstraints)

setDiet(theObject, dietf)

setDiet.name(theObject, diet_name)

writeFBAfile(theObject, fba_fname, ...)
}
\arguments{
\item{theObject}{A \code{FBA_greatmod} object}

\item{reaction.name}{Character of the reaction name.}

\item{newConstraints}{Numeric vector of length equal to 2 containing the new lower and upper constraints}

\item{dietf}{The diet file or data.frame}

\item{diet_name}{Character for the diet name}

\item{fba_fname}{The file name in which the FBA model will be saved. See \emph{epimod::model.generation}.}
}
\value{
The constraints of the reaction.

The constraints of the reaction.

The FBA_greatmod class with the the constraints updated.

The FBA_greatmod class with lower bounds updated with diet values.

The FBA_greatmod class with the diet name updated.
}
\description{
Write the fba model in file.
}
