% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{process_model}
\alias{process_model}
\title{Process a single biounit model}
\usage{
process_model(m, hypernode_name, mat_dir, base_dir = getwd())
}
\arguments{
\item{m}{A list for one cellular unit (from
\code{make_biounit_models()}).}

\item{hypernode_name}{Character. Parent hyper-node label.}

\item{mat_dir}{Directory that contains the user-supplied
\emph{.mat} files (same value you pass to
\code{build_hypernode()}).}

\item{base_dir}{Root working directory (default = \code{getwd()}).}
}
\value{
(Invisibly) a list with status and created paths.
}
\description{
\itemize{
\item Copies the chosen \emph{.mat} file into the hyper-node workspace
\item Converts it to a text FBA model with \code{writeFBAfile()}
\item Returns the path to that single \emph{.txt} file
}
}
