% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_boundary_reactions.R
\name{project_boundary_reactions}
\alias{project_boundary_reactions}
\title{Project boundary reactions across a set of FBA models}
\usage{
project_boundary_reactions(
  biounit_models,
  boundary_metabolites,
  out_dir,
  hypernode_name
)
}
\description{
Project boundary reactions across a set of FBA models
}
